feat(interview): 添加 STT 使用记录功能
- 新增记录 STT 使用情况的接口和相关逻辑
- 实现分布式锁机制，确保并发情况下数据一致性
- 添加使用记录表和相关 mapper
- 优化面试信息表结构，增加 STT 和 AG 使用统计字段

+++ src/main/java/com/keji/green/lit/engine/mapper/InterviewInfoMapper.java
@@ -44,7 +44,7 @@ public interface InterviewInfoMapper {
      * @param costInCents 语音识别花费
      * @return 更新记录数
      */
-    int updateSttUsageByInterviewId(String interviewId, Long durationSeconds, Integer costInCents);
+    int updateSttUsageByInterviewId(String interviewId, Long durationSeconds, Long costInCents);
 
     /**
      * 更新面试的语音识别时长和语音识别花费
@@ -52,5 +52,5 @@ public interface InterviewInfoMapper {
      * @param costInCents 语音识别花费
      * @return 更新记录数
      */
-    int updateAgUsageByInterviewId(String interviewId, Integer costInCents);
+    int updateAgUsageByInterviewId(String interviewId, Long costInCents);
 }
